Use the MIDI status byte to determine keyDown
The MIDI status byte is present in _every_ MIDI message, and indicates
what the message actually is.
For a keyDown message, the byte is 0x90.

Many keyboards emit non-zero velocities on key up, and some are even
variable based on the key return speed.

This commit changes both the MIDI message processor to _only_ look at
messages with a keycode of 0x80 (keyUp) or 0x90 (keyDown), as well as
ensuring the previous note range clamp is in effect. It fixes #341.

Furthermore, it fixes a class of issues where other control surface
inputs were being captured as notes.