Restart midi input on port change
In the current implementation, if you start a MIDI input session, and
then change the device/port, your existing MIDI session will continue
with the old device/port.

This ends that session and immediately starts a new one, effectively
"restarting it."

JZZ was unclear if their package allows for changing a port while a read
callback is running. Some other WebMIDI systems, such as webmidi.js, do
allow for this.