Handle keyUp status in running status mode
According to the MIDI spec, a keyDown with a velocity of 0 is equivalent
to a keyUp status.
This is how the code previously worked, prior to 71300a0.

This commit reintroduces a check for that, while preserving the newer
(as of 71300a0) status checks. That way, both keyUp AND keyDowns with 0
velocity are registered as keyUps